Move ecs-service class into service package
