+++ src/main/java/com/emc/ecs/serviceBroker/service/EcsServiceInstanceBindingService.java
@@ -22,7 +22,6 @@ import org.springframework.stereotype.Service;
 import com.emc.ecs.managementClient.model.UserSecretKey;
 import com.emc.ecs.serviceBroker.EcsManagementClientException;
 import com.emc.ecs.serviceBroker.EcsManagementResourceNotFoundException;
-import com.emc.ecs.serviceBroker.EcsService;
 import com.emc.ecs.serviceBroker.repository.ServiceInstanceBinding;
 import com.emc.ecs.serviceBroker.repository.ServiceInstanceBindingRepository;
