+++ src/test/java/com/emc/ecs/serviceBroker/repository/ServiceInstanceRepositoryTest.java
@@ -20,8 +20,8 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import com.emc.ecs.serviceBroker.EcsManagementClientException;
 import com.emc.ecs.serviceBroker.EcsManagementResourceNotFoundException;
-import com.emc.ecs.serviceBroker.EcsService;
 import com.emc.ecs.serviceBroker.config.Application;
+import com.emc.ecs.serviceBroker.service.EcsService;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = Application.class,
